+++ examples/ssr-islands-axum/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "leptos-fluent-ssr-hydrate-axum-islands-example"
+name = "leptos-fluent-ssr-islands-axum-example"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -21,7 +21,6 @@ tower = { version = "0.4", optional = true, features = ["util"] }
 tower-http = { version = "0.5", features = ["fs"], optional = true }
 wasm-bindgen = "=0.2.93"
 thiserror = "1"
-tracing = { version = "0.1", optional = true }
 http = "1"
 
 leptos-fluent = { path = "../../leptos-fluent" }
@@ -46,7 +45,6 @@ ssr = [
   "leptos/ssr",
   "leptos_meta/ssr",
   "leptos_router/ssr",
-  "dep:tracing",
   "leptos-fluent/ssr",
   "leptos-fluent/axum",
 ]
