--- examples/ssr-islands-axum/locales/core/en/main.ftl
@@ -1,1 +1,0 @@
-slogan = This is my slogan
