--- examples/ssr-islands-axum/locales/core/es/main.ftl
@@ -1,1 +1,0 @@
-slogan = Este es mi slogan
